Added custom resolution methods to @Bind
If the "uses" parameter is present, the scanner will generate a
Route::bind instead of Route::model.  The resolution method is specified
in "action" format: "FQCN@method".  If a method isn't specified, the
default "bind" method is called on the specified class.
E.g.: @Bind("user", uses="App\User@findBySlug")